Separate SDL Rendering into it's own class
move SDL Rendering to SDLRendering class
